PreparedStatus redesign for GetStatusHandler
